+++ src/main/java/dev/orne/qdsl/EntityPathTranslator.java
@@ -24,6 +24,9 @@ package dev.orne.qdsl;
 
 import javax.validation.constraints.NotNull;
 
+import org.apache.commons.lang3.Validate;
+
+import com.querydsl.core.support.ReplaceVisitor;
 import com.querydsl.core.types.EntityPath;
 import com.querydsl.core.types.Expression;
 import com.querydsl.core.types.FactoryExpression;
@@ -41,10 +44,39 @@ import com.querydsl.core.types.ProjectionRole;
  * @since 0.1
  */
 public class EntityPathTranslator<S, I>
-extends AbstractPathTranslator<S, I> {
+extends ReplaceVisitor<Void>
+implements ValueAssignmentReplaceVisitor<Void> {
 
+    /** The source path. */
+    private final @NotNull EntityPath<S> source;
+    /** The source path type. */
+    private final @NotNull Class<? extends S> sourceType;
+    /** The target expression. */
+    private final @NotNull Path<I> target;
     /** The projection expression. */
     private final @NotNull Expression<S> projection;
+    /** The value assignment translator. */
+    private final @NotNull AssignmentTranslator<S> assignmentTranslator;
+
+    /**
+     * Creates a new instance.
+     * 
+     * @param source The source path
+     * @param target The target expression
+     * @param assignmentTranslator The value assignment translator
+     */
+    protected EntityPathTranslator(
+            final @NotNull EntityPath<S> source,
+            final @NotNull Path<I> target,
+            final @NotNull Expression<S> projection,
+            final @NotNull AssignmentTranslator<S> assignmentTranslator) {
+        super();
+        this.source = Validate.notNull(source);
+        this.sourceType = source.getType();
+        this.target = Validate.notNull(target);
+        this.projection = Validate.notNull(projection);
+        this.assignmentTranslator = Validate.notNull(assignmentTranslator);
+    }
 
     /**
      * Builder constructor.
@@ -53,8 +85,13 @@ extends AbstractPathTranslator<S, I> {
      */
     protected EntityPathTranslator(
             final @NotNull Builder<S, I> builder) {
-        super(builder);
+        super();
+        Validate.notNull(builder);
+        this.source = builder.source;
+        this.sourceType = this.source.getType();
+        this.target = builder.target;
         this.projection = builder.projection;
+        this.assignmentTranslator = builder.assignmentTranslator;
     }
 
     public static <V> TargetBuilder<V> fromPath(
@@ -63,23 +100,30 @@ extends AbstractPathTranslator<S, I> {
     }
 
     /**
-     * Returns the source entity path.
+     * Returns the source path.
      * 
-     * @return The source entity path
+     * @return The source path
      */
-    @Override
     public @NotNull EntityPath<S> getSource() {
-        return (EntityPath<S>) super.getSource();
+        return this.source;
     }
 
     /**
-     * Returns the target identity path.
+     * Returns the source path type.
      * 
-     * @return The target identity path
+     * @return The source path type
      */
-    @Override
+    public @NotNull Class<? extends S> getSourceType() {
+        return this.sourceType;
+    }
+
+    /**
+     * Returns the target expression.
+     * 
+     * @return The target expression
+     */
     public @NotNull Path<I> getTarget() {
-        return (Path<I>) super.getTarget();
+        return this.target;
     }
 
     /**
@@ -89,6 +133,47 @@ extends AbstractPathTranslator<S, I> {
      */
     public Expression<S> getProjection() {
         return this.projection;
+    }
+
+    /**
+     * Checks if the expression is the source path.
+     * 
+     * @param expr The expression to check
+     * @return If the expression is the source path
+     */
+    protected boolean isSource(
+            final @NotNull Expression<?> expr) {
+        return this.source.equals(expr);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Expression<?> visit(
+            final Path<?> expr,
+            final Void context) {
+        if (isSource(expr))  {
+            return this.target;
+        } else {
+            return super.visit(expr, context);
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public @NotNull ValueAssignments visit(
+            final @NotNull ValueAssignment<?> vexpr,
+            final Void context) {
+        if (isSource(vexpr.getPath())) {
+            @SuppressWarnings("unchecked")
+            final Expression<S> value = (Expression<S>) vexpr.getValue();
+            return this.assignmentTranslator.apply(value);
+        } else {
+            return ValueAssignments.of(vexpr);
+        }
     }
 
     /**
@@ -222,9 +307,14 @@ extends AbstractPathTranslator<S, I> {
      * @since 0.1
      */
     protected static class Builder<S, I>
-    extends AbstractPathTranslator.Builder<S, I>
     implements ProjectionBuilder<S, I>, StoreBuilder<S, I>, FinalBuilder<S, I> {
 
+        /** The source path. */
+        private final @NotNull EntityPath<S> source;
+        /** The target expression. */
+        private Path<I> target;
+        /** The value assignment translator. */
+        private AssignmentTranslator<S> assignmentTranslator;
         /** The projection expression. */
         private Expression<S> projection;
 
@@ -235,10 +325,58 @@ extends AbstractPathTranslator<S, I> {
          * @param identity The identity path
          */
         protected Builder(
-                final @NotNull Path<S> source,
+                final @NotNull EntityPath<S> source,
                 final @NotNull Path<I> identity) {
-            super(source);
+            super();
+            this.source = Validate.notNull(source);
             setTarget(identity);
+        }
+
+        /**
+         * Returns the source path.
+         * 
+         * @return The source path
+         */
+        protected EntityPath<S> getSource() {
+            return this.source;
+        }
+
+        /**
+         * Returns the target expression.
+         * 
+         * @return The target expression
+         */
+        protected Path<I> getTarget() {
+            return this.target;
+        }
+
+        /**
+         * Sets the target expression.
+         * 
+         * @param target The target expression
+         */
+        protected void setTarget(
+                final @NotNull Path<I> target) {
+            this.target = target;
+        }
+
+        /**
+         * Returns the value assignment translator.
+         * 
+         * @return The value assignment translator
+         */
+        protected AssignmentTranslator<S> getAssignmentTranslator() {
+            return this.assignmentTranslator;
+        }
+
+        /**
+         * Sets the value assignment translator.
+         * 
+         * @param translator The value assignment translator
+         */
+        protected void setAssignmentTranslator(
+                final @NotNull AssignmentTranslator<S> translator) {
+            this.assignmentTranslator = translator;
         }
 
         /**
@@ -288,7 +426,7 @@ extends AbstractPathTranslator<S, I> {
                 final @NotNull ExpressionTranslator<S, I> exprTranslator,
                 final @NotNull ValueTranslator<S, I> valueTranslator) {
             setAssignmentTranslator(AssignmentTranslator.forPath(
-                    (Path<I>) super.getTarget(),
+                    this.target,
                     valueTranslator,
                     exprTranslator));
             return this;